rm: fix '-r' flag
The '-r' flag was implmented wrong, rewrote the function using recursion
to loop over dir contents. The check for '.' and '..' is unneeded given
the node.js specification of 'fs.readdir'

`rm -rf /` seem to be failing due to missing '.deletedFiles.log' after
first iterations of the function.

Removed unnecessary comment in xargs