node/lib/buffer: fix Buffer() when called w/ ArrayBuffer (Firefox 48)
I don't really know why, but Firefox 48 started doing this, and our
tests broke.  this is based off of what node v6 does, but modified to
work with minimal changes to the rest of buffer.js.  Additionally - we
can't call allocate(), as this results in a stack overflow, so ignore
the pool and just create a new Buffer object the old-fashioned way.