kernel: special case opening '/dev/null', fix general open() case
the call to addFile must happen in general in the callback from
fs.open(), or else we aren't guaranteed to have a valid file object in
f.  In tests/many cases this might Just Work, if open() is able to
immediately complete (like for the in-memory filesystem) and the
callback is called before kernel.fs.open() returns, but we can't count
on that.